Huge refactor of Slide, first step of intergrationg instuctor review (UNSTABLE)
includes:
code style for interfaces
new actions/api for redux
refactor of slide
extracting slide rendered, slide context, render context

(UNSTABLE) due to integration problems, insructor review will throw an error on render in dev server

+++ src/Frontend/src/api/groups.ts
@@ -1,92 +1,113 @@
 import api from "src/api/index";
-import { resetStudentsLimits } from "src/consts/routes";
+import { groups, resetStudentsLimits } from "src/consts/routes";
 import {
 	CopyGroupResponse,
 	GroupAccessesResponse,
 	GroupInfo,
-	GroupScoringGroupsResponse,
+	GroupScoringGroupsResponse, GroupsInfoResponse,
 	GroupStudentsResponse
 } from "src/models/groups";
+import { buildQuery } from "src/utils";
+import { Dispatch } from "redux";
+import { groupLoadFailAction, groupLoadStartAction, groupLoadSuccessAction } from "src/actions/groups";
 
 // Groups
-export function getCourseGroups(courseId: string): Promise<{ groups: GroupInfo[] }> {
-	return api.get("groups?courseId=" + courseId);
+export function getCourseGroups(courseId: string, userId?: string,): Promise<{ groups: GroupInfo[] }> {
+	const url = groups + buildQuery({ courseId, userId, });
+	return api.get(url);
+}
+
+export function getCourseGroupsRedux(courseId: string, userId: string,) {
+	return (dispatch: Dispatch): Promise<GroupsInfoResponse | string> => {
+		dispatch(groupLoadStartAction(courseId, userId));
+		const url = groups + buildQuery({ courseId, userId, });
+		return api.get<GroupsInfoResponse>(url)
+			.then(json => {
+				dispatch(groupLoadSuccessAction(courseId, userId, json));
+				return json;
+			})
+			.catch(error => {
+				dispatch(groupLoadFailAction(courseId, userId, error));
+				return error;
+			});
+	};
 }
 
 export function getCourseArchivedGroups(courseId: string): Promise<{ groups: GroupInfo[] }> {
-	return api.get("groups?courseId=" + courseId + "&archived=true");
+	const url = groups + buildQuery({ courseId, archived: true, });
+	return api.get(url);
 }
 
 // Group
 export function getGroup(groupId: number): Promise<GroupInfo> {
-	return api.get("groups/" + groupId);
+	return api.get(`${ groups }/${ groupId }`);
 }
 
 export function createGroup(courseId: string, name: string): Promise<Response> {
-	return api.post("groups?courseId=" + courseId,
-		api.createRequestParams({ name }));
+	const url = groups + buildQuery({ courseId });
+	return api.post(url, api.createRequestParams({ name }));
 }
 
 export function copyGroup(groupId: number, destinationCourseId: string,
 	makeMeOwner: boolean
 ): Promise<CopyGroupResponse> {
-	return api.post("groups/" + groupId + "/copy?destinationCourseId="
-		+ encodeURIComponent(destinationCourseId) + '&makeMeOwner=' + makeMeOwner);
+	const url = `${ groups }/${ groupId }/copy` + buildQuery({ destinationCourseId, makeMeOwner, });
+	return api.post(url);
 }
 
 export function saveGroupSettings(groupId: number, groupSettings: Record<string, unknown>): Promise<GroupInfo> {
-	return api.patch("groups/" + groupId,
+	return api.patch(`${ groups }/${ groupId }`,
 		api.createRequestParams(groupSettings));
 }
 
 export function deleteGroup(groupId: number): Promise<Response> {
-	return api.delete("groups/" + groupId);
+	return api.delete(`${ groups }/${ groupId }`);
 }
 
 export function changeGroupOwner(groupId: number, ownerId: string): Promise<Response> {
-	return api.put("groups/" + groupId + '/owner',
+	return api.put(`${ groups }/${ groupId }/owner`,
 		api.createRequestParams({ ownerId }));
 }
 
 // Scores
 export function getGroupScores(groupId: number): Promise<GroupScoringGroupsResponse> {
-	return api.get("groups/" + groupId + '/scores');
+	return api.get(`${ groups }/${ groupId }/scores`);
 }
 
 export function saveScoresSettings(groupId: number, checkedScoresSettingsIds: string[]): Promise<Response> {
-	return api.post("groups/" + groupId + '/scores',
+	return api.post(`${ groups }/${ groupId }/scores`,
 		api.createRequestParams({ 'scores': checkedScoresSettingsIds }));
 }
 
 // Accesses
 export function getGroupAccesses(groupId: number): Promise<GroupAccessesResponse> {
-	return api.get("groups/" + groupId + "/accesses");
+	return api.get(`${ groups }/${ groupId }/accesses`);
 }
 
 export function addGroupAccesses(groupId: number, userId: string): Promise<Response> {
-	return api.post("groups/" + groupId + "/accesses/" + userId);
+	return api.post(`${ groups }/${ groupId }/accesses/${ userId }`);
 }
 
 export function removeAccess(groupId: number, userId: string): Promise<Response> {
-	return api.delete("groups/" + groupId + "/accesses/" + userId);
+	return api.delete(`${ groups }/${ groupId }/accesses/${ userId }`);
 }
 
 // Students
 export function getStudents(groupId: number): Promise<GroupStudentsResponse> {
-	return api.get("groups/" + groupId + "/students");
+	return api.get(`${ groups }/${ groupId }/students`);
 }
 
 export function deleteStudents(groupId: number, studentIds: string[]): Promise<Response> {
-	return api.delete("groups/" + groupId + "/students",
+	return api.delete(`${ groups }/${ groupId }/students`,
 		api.createRequestParams({ studentIds }));
 }
 
 export function copyStudents(groupId: number, studentIds: string[]): Promise<Response> {
-	return api.post("groups/" + groupId + "/students",
+	return api.post(`${ groups }/${ groupId }/students`,
 		api.createRequestParams({ studentIds }));
 }
 
 export function resetLimitsForStudents(groupId: number, studentIds: string[]): Promise<Response> {
-	return api.post("groups/" + groupId + resetStudentsLimits,
+	return api.post(`${ groups }/${ groupId }/${ resetStudentsLimits }`,
 		api.createRequestParams({ studentIds }));
 }
